app/requirements.txt
Flask
spacy
sentence-transformers
pdfplumber
python-docx
pandas
notebook